+++ src/senaite/sync/browser/views.py
@@ -65,13 +65,23 @@ class Sync(BrowserView):
             url = storage["credentials"]["url"]
             username = storage["credentials"]["username"]
             password = storage["credentials"]["password"]
-            content_types = storage["configuration"].get("content_types", None)
+            prefix = self.get_storage_config(domain_name, "prefix", None)
+            content_types = self.get_storage_config(
+                                    domain_name, "content_types", [])
+            unwanted_content_types = self.get_storage_config(
+                                    domain_name, "unwanted_content_types", [])
+            prefixable_types = self.get_storage_config(
+                                    domain_name, "prefixable_types", [])
+
             data = {
                 "url": url,
                 "domain_name": domain_name,
                 "ac_name": username,
                 "ac_password": password,
                 "content_types": content_types,
+                "unwanted_content_types": unwanted_content_types,
+                "prefix": prefix,
+                "prefixable_types": prefixable_types,
             }
             step = ImportStep(data)
             step.run()
@@ -100,7 +110,14 @@ class Sync(BrowserView):
             url = storage["credentials"]["url"]
             username = storage["credentials"]["username"]
             password = storage["credentials"]["password"]
-            content_types = storage["configuration"].get("content_types", None)
+            prefix = self.get_storage_config(domain_name, "prefix", None)
+            content_types = self.get_storage_config(
+                                    domain_name, "content_types", [])
+            unwanted_content_types = self.get_storage_config(
+                                    domain_name, "unwanted_content_types", [])
+            prefixable_types = self.get_storage_config(
+                                    domain_name, "prefixable_types", [])
+
             data = {
                 "url": url,
                 "domain_name": domain_name,
@@ -108,6 +125,9 @@ class Sync(BrowserView):
                 "ac_password": password,
                 "fetch_time": fetch_time,
                 "content_types": content_types,
+                "unwanted_content_types": unwanted_content_types,
+                "prefix": prefix,
+                "prefixable_types": prefixable_types,
             }
             step = ComplementStep(data)
             step.run()
@@ -122,52 +142,17 @@ class Sync(BrowserView):
             self.add_status_message(message, "info")
             return self.template()
 
-        # Handle "Fetch" action
-        if form.get("fetch", False):
-
-            url = form.get("url", "")
-            if not url.startswith("http"):
-                url = "http://{}".format(url)
-            domain_name = form.get("domain_name", None)
-            username = form.get("ac_name", None)
-            password = form.get("ac_password", None)
-            # check if all mandatory fields have values
-            if not all([domain_name, url, username, password]):
-                message = _("Please fill in all required fields")
-                self.add_status_message(message, "error")
-                return self.template()
-
-            import_settings = True if form.get("import_settings") == 'on' else False
-            import_users = True if form.get("import_users") == 'on' else False
-            import_registry = True if form.get("import_registry") == 'on' else False
-            content_types = form.get("content_types", None)
-            if content_types is not None:
-                content_types = [t.strip() for t in content_types.split(",")]
-                portal_types = api.get_tool("portal_types")
-                content_types = filter(lambda ct: ct in portal_types,
-                                       content_types)
-
-            data = {
-                "url": url,
-                "domain_name": domain_name,
-                "ac_name": username,
-                "ac_password": password,
-                "content_types": content_types,
-                "import_settings": import_settings,
-                "import_users": import_users,
-                "import_registry": import_registry,
-            }
-
-            fs = FetchStep(data)
-            verified, message = fs.verify()
-            if verified:
-                fs.run()
-                self.add_status_message(message, "info")
-            else:
-                self.add_status_message(message, "error")
-
         # always render the template
         return self.template()
+
+    def get_storage_config(self, domain_name, config_name, default = None):
+        """ Get the advanced configuration setting for a given domain
+        :param config_name: advanced configuration section name
+        :param default: default value if configuration value is not set
+        :return:
+        """
+        storage = self.get_storage(domain_name)
+        return storage["configuration"].get(config_name, default)
 
     def add_status_message(self, message, level="info"):
         """Set a portal status message