+++ src/senaite/sync/fetchstep.py
@@ -64,12 +64,15 @@ class FetchStep(SyncStep):
         storage["credentials"]["password"] = self.password
         # remember import configuration in the storage
         storage["configuration"]["content_types"] = self.content_types
+        storage["configuration"]["unwanted_content_types"] = self.unwanted_content_types
+        storage["configuration"]["prefix"] = self.prefix
+        storage["configuration"]["prefixable_types"] = self.prefixable_types
         storage["configuration"]["import_settings"] = self.import_settings
         storage["configuration"]["import_registry"] = self.import_registry
         storage["configuration"]["import_users"] = self.import_users
         storage["last_fetch_time"] = DateTime()
 
-        message = "Fetching Data started for {}".format(self.domain_name)
+        message = "Data fetched and saved: {}".format(self.domain_name)
         return True, message
 
     def get_version(self):
@@ -123,7 +126,7 @@ class FetchStep(SyncStep):
                     start_from, start_from+window))
             for item in items:
                 # skip object or extract the required data for the import
-                if not item or not item.get("portal_type", True):
+                if not self.is_item_allowed(item):
                     continue
                 data_dict = utils.get_soup_format(item)
                 rec_id = self.sh.insert(data_dict)