+++ src/senaite/sync/utils.py
@@ -7,11 +7,12 @@ from senaite import api
 from senaite.sync import logger
 from DateTime import DateTime
 from datetime import datetime
+from senaite.sync.souphandler import REMOTE_UID, REMOTE_PATH, PORTAL_TYPE
 
 
-SOUPER_REQUIRED_FIELDS = {"uid": "remote_uid",
-                          "path": "path",
-                          "portal_type": "portal_type"}
+SOUPER_REQUIRED_FIELDS = {"uid": REMOTE_UID,
+                          "path": REMOTE_PATH,
+                          "portal_type": PORTAL_TYPE}
 
 SYNC_CREDENTIALS = "senaite.sync.credentials"
 
@@ -30,19 +31,39 @@ def to_review_history_format(review_history):
     return review_history
 
 
-def is_item_allowed(item):
+def has_valid_portal_type(item):
     """ Check if an item can be handled based on its portal type.
     :return: True if the item can be handled
     """
     if not isinstance(item, dict):
         return False
 
-    portal_types = api.get_tool("portal_types")
+    portal_types = api.get_tool("portal_types").listContentTypes()
     pt = item.get("portal_type", None)
     if pt not in portal_types:
         return False
 
     return True
+
+
+def filter_content_types(content_types):
+    """
+
+    :param content_types:
+    :return:
+    """
+    ret = list()
+    if not content_types:
+        return ret
+
+    # Get available portal types and make it all lowercase
+    portal_types = api.get_tool("portal_types").listContentTypes()
+    portal_types = [t.lower for t in portal_types]
+
+    ret = [t.strip() for t in content_types.split(",") if t]
+    ret = filter(lambda ct, types=portal_types: ct.lower() in types, ret)
+    ret = list(set(ret))
+    return ret
 
 
 def get_parent_path(path):
@@ -56,7 +77,9 @@ def get_parent_path(path):
     if path.endswith("/"):
         path = path[:-1]
     parts = path.split("/")
-    return "/".join(parts[:-1])
+    if len(parts) < 3:
+        return "/"
+    return str("/".join(parts[:-1]))
 
 
 def get_id_from_path(path):